+++ CHANGELOG.md
@@ -1,3 +1,29 @@
+## [4.21.1](https://github.com/spaceship-prompt/spaceship-prompt/compare/v4.21.0...v4.21.1) (2026-03-02)
+
+
+### Bug Fixes
+
+* initialize async worker before chpwd hook uses it ([#1527](https://github.com/spaceship-prompt/spaceship-prompt/issues/1527)) ([3043f47](https://github.com/spaceship-prompt/spaceship-prompt/commit/3043f478941173f1af8da8d5179a0869f9f05005))
+
+# [4.21.0](https://github.com/spaceship-prompt/spaceship-prompt/compare/v4.20.0...v4.21.0) (2025-12-24)
+
+
+### Features
+
+* **haxe:** add support for haxe ([#1513](https://github.com/spaceship-prompt/spaceship-prompt/issues/1513)) ([ff0d638](https://github.com/spaceship-prompt/spaceship-prompt/commit/ff0d638cf2ebeb07490b3205f853ddb8e6c13fe6))
+
+# [4.20.0](https://github.com/spaceship-prompt/spaceship-prompt/compare/v4.19.3...v4.20.0) (2025-12-20)
+
+
+### Bug Fixes
+
+* make rlang.test.zsh executable ([bdb073c](https://github.com/spaceship-prompt/spaceship-prompt/commit/bdb073c50964a58b6448212beddf1c1f5c2c5d67))
+
+
+### Features
+
+* **rlang:** implement new section to support RLang ([3361bb7](https://github.com/spaceship-prompt/spaceship-prompt/commit/3361bb748a26305b25159cb63f2395b7e785e35a))
+
 ## [4.19.3](https://github.com/spaceship-prompt/spaceship-prompt/compare/v4.19.2...v4.19.3) (2025-12-19)